+++ CHANGELOG.rst
@@ -9,3 +9,4 @@ Unreleased
 * Changed: WorkbenchParser.parse to log parse error instead of raising an error. Error is usually due to server sending
 malformed XML instead of an actual erroneous condition.
 * Added: Support of Container Security Containers API.
+* Added: Support of Container Security Test Jobs API.